+++ backend/drm/util.c
@@ -9,7 +9,7 @@
 #include "backend/drm/util.h"
 
 int32_t calculate_refresh_rate(const drmModeModeInfo *mode) {
-	int32_t refresh = (mode->clock * 1000000LL / mode->htotal +
+	int32_t refresh = (((int64_t) mode->clock) * 1000000LL / mode->htotal +
 		mode->vtotal / 2) / mode->vtotal;
 
 	if (mode->flags & DRM_MODE_FLAG_INTERLACE) {
